Update APT repository for Docker engine
Docker recently updated their package repositories. These changes add
support for the new repositories to this role.

For more details on the change, see: http://blog.docker.com/2015/07/new-apt-and-yum-repos/

If you're running this on a system that was using an older version of
this role, please ensure that you remove the older Docker package name
first:

```
$ sudo apt-get purge lxc-docker*
```